Change API and add error checking to addTromino
The API change is required to avoid throwing exceptions for
non-exceptional behaviour. This behaviour (not being able to add the
tromino because the position is invalid) was not checked before this
commit.

tile(Position, Position) no longer calls addTromino because tile
modifies existing state and addTromino no longer does. This will be
rectified in a future commit, providing a non mutable API for tile,
whilst using the mutable version internally.

The reason tile mutates state is simplicity of the code, and this will
remain despite the future API changes.