API change for View, Canvas, and Main. Some functionality altered
The new API allows for dependency injection and easier testing.

Some functionality of the GUI has changed, specifically:
Randomness is no longer present in the Canvas generation, and nor is an
automatic board.tile call.
The border no longer exists (there is a small border in the y direction,
but this is unintentional) - this was present when the larger border was
in place but less noticable.
The canvas and view are now resizable, although there exists a bug with
snapping to the border. This is probably related to the y axis border.